requirements.txt
google-adk[eval]==1.18.0
chromadb==1.3.5